move the changeStage metho to LeadModel
